Cambios hasta el 7/06/2023

+++ app/Models/Persona.php
@@ -50,7 +50,6 @@ class Persona extends Model
 
             return $persona;
         } else if ($evento == 'U') {
-
             $persona = Persona::find($request->persona_id);
             $persona->numero_identificacion = $request->get('numero_identificacion');
             $persona->grado = $request->get('grado')== 0 ? null : $request->get('grado');
@@ -62,25 +61,29 @@ class Persona extends Model
             $persona->cargo = $request->get('cargo');
             $persona->usuario_modificador = $request->get('usuario');
             $persona->fecha_modificacion = DB::raw('GETDATE()');
-            $persona->save();
-
+            $folderPath = public_path() . '/img/perfil';
+            $foto = $folderPath . '/' . $request->get('tmp_numero_identificacion') . $request->get('tipo_imagen');
+            
             if ($request->get('imagen') != null) {
-                $folderPath = public_path() . '/img/perfil';
                 if (!File::exists($folderPath)) {
                    File::makeDirectory($folderPath, 0755, true);
                 }
             
-                $foto = $folderPath . '/' . $persona->numero_identificacion . $request->get('tipo_imagen');
-                            
-                if (file_exists($foto)) {
-                  File::delete($foto);
+                if ($request->get('tmp_imagen') != null) {
+                    $foto = $folderPath . '/' . $request->get('tmp_imagen');
+
+                    if (file_exists($foto)) {
+                        File::delete($foto);
+                    }
                 }
                     
                 $image_base64 = base64_decode($request->get('imagen'));
                 file_put_contents($foto, $image_base64);
             
-                $persona->imagen = $persona->Cedula . $request->get('tipo_imagen');
+                $persona->imagen = $persona->numero_identificacion . $request->get('tipo_imagen');
             }
+
+            $persona->save();
 
             return $persona;
         }
